fix: store `price` in BigDecimal

+++ backend/src/main/java/com/bugenzhao/bookstore_backend/entity/db/Book.java
@@ -1,17 +1,19 @@
 package com.bugenzhao.bookstore_backend.entity.db;
 
+import java.math.BigDecimal;
+
 public class Book {
     public int id;
     public String isbn;
     public String name;
     public String type;
     public String author;
-    public double price;
+    public BigDecimal price;
     public String description;
     public int inventory;
     public String image;
 
-    public Book(int id, String isbn, String name, String type, String author, double price, String description,
+    public Book(int id, String isbn, String name, String type, String author, BigDecimal price, String description,
             int inventory, String image) {
         this.id = id;
         this.isbn = isbn;
